requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
google-generativeai==0.3.2
firebase-admin==6.2.0
python-dotenv==1.0.0
aiosqlite==0.19.0
requests==2.31.0
pytz==2023.3
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
Pillow
huggingface-hub==0.19.4